AY-52-Added Service Cards and restructure overall layout
