+++ web/index.html
@@ -4,10 +4,10 @@
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/vite.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Vite + React + TS</title>
+    <title>Hands Off</title>
   </head>
   <body class="h-full w-full">
-    <div id="root" class="h-full w-full"></div>
+    <div id="root" class="h-full w-full max-w-screen p-0 m-0"></div>
     <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
