+++ web/src/App.tsx
@@ -5,16 +5,13 @@ import Navbar from "./components/Navbar";
 import ServiceDetailsPage from "./pages/ServiceDetails";
 import SignUpPage from "./pages/SignUp";
 import {
-  BrowserRouter as Router,
+  BrowserRouter,
   Routes,
   Route,
-  BrowserRouter,
+ 
 } from "react-router-dom";
+import type { Link } from "./Types";
 
-interface Link {
-  url: string;
-  label: string;
-}
 
 function App() {
   const links: Link[] = [
@@ -29,10 +26,12 @@ function App() {
       <div className="flex w-full justify-end">
         <Navbar links={links} isLogged />
       </div>
+      <div className="w-full h-full p-4">
       <Routes>
         <Route path="/services" element={<ServiceDetailsPage />} />
         <Route path="/sign-up" element={<SignUpPage />} />
       </Routes>
+      </div>
     </BrowserRouter>
   );
 }
