+++ web/src/Types.tsx
@@ -1,0 +1,13 @@
+interface serviceType {
+  name: string;
+  price: number;
+  description: string;
+  img_url: string;
+}
+interface Link {
+  url: string;
+  label: string;
+}
+
+
+export type {serviceType, Link};
